+++ modules/types/src/node.ts
@@ -81,6 +81,10 @@ export interface INodeService {
     params: OptionalPublicIdentifier<NodeParams.SignUtilityMessage>,
   ): Promise<Result<NodeResponses.SignUtilityMessage, NodeError>>;
 
+  sendIsAliveMessage(
+    params: OptionalPublicIdentifier<NodeParams.SendIsAlive>,
+  ): Promise<Result<NodeResponses.SendIsAlive, NodeError>>;
+
   // Dispute methods
   sendDisputeChannelTx(
     params: OptionalPublicIdentifier<NodeParams.SendDisputeChannelTx>,