+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,15 @@
 
 ## Next Release
 
+## 0.1.0-rc.1
+
+- \[router\] add isAlive check before forwarding transfer to increase odds of successful transfer
+## 0.1.0-rc.0
+
+- \[contracts\] audit results implemented
+- \[router\] store and process queued updates
+- \[engine\] restore state from counterparty
+- \[engine\] handle `isAlive` messages
 - \[router\] fix rebalance profile getter to respect assetIds
 - \[router\] cancel transfers if they are unsuccessful
 - \[contracts\] added required `EncodedCancel` getter to transfer registry interface